Initial commit of prototype Electron app
Using Electron 1.7.9 and a very basic hand-crafted interface. Not yet using
React or Redux, but will soon. For now, it fetches the tags from the backend
and shows them (unformatted) on the page. A drop zone is defined that
receives files and uploads them to the backend, showing the result in the
page.