Remove electron-redux as it is not helpful
When combining React and Redux, synchronizing the state between processes in
Electron is detrimental. As such, remove this module. Each process must keep
its own state and any synchronization will be done using IPC. For now, there
is no need for that, as the state is all about the interface in a particular
browser window anyway.

npm test passes