Allow dropping files while on the uploads page
The uploads page can now receive additional dropped files. It also ignores
files that already appear on the uploads page. The form data is also left
untouched when new files are dropped on the uploads page.

Fixed a couple of errors that only appeared in the console.

npm test passes