Implement the search facility
Search is not entirely working, in that after: and before: are causing the
search results to come back empty. Otherwise the functionality is in place
and basically working.

Fixed the redux logging so it does not show during tests.

Added the checksum algorithm prefix to the checksum values (for upload).

npm test passes