Show details, not filters, for asset page
It makes no sense to show the asset filters on the left when showing the
asset details page. Instead, show the asset sidebar, with all of the
details, including tags and such. These details are no longer on the asset
preview component. Also, the asset preview image scales to fit the contents,
just like the edit page, which actually looks better than not stretching.

npm test passes