Move things around so electron-builder works well
Rename 'app' to 'src' and compile everything to the 'app' directory. Fix
package.json so it has a main that points to the app/main.js file. No need
to have a separate package.json in the app directory any more.

This allows the introduction of electron-builder for building a macOS pkg
for installing on macOS systems. Tested successfully.

npm test passes